+++ src/color-picker/ColorPicker.stories.tsx
@@ -6,10 +6,9 @@ const ExampleColorPicker = ({ color }: { color: string }) => {
 	const [value, setValue] = useState(color);
 
 	return (
-		<Rangu.Root
+		<Rangu
 			value={value}
-			onChange={setValue}
-			outputFormat="rgba"
+			onChange={({ rgba }) => setValue(rgba)}
 		>
 			<div
 				style={{
@@ -54,7 +53,10 @@ const ExampleColorPicker = ({ color }: { color: string }) => {
 				</div>
 
 				<Rangu.EyeDropper />
-				<Rangu.InputFields withLabels />
+				<Rangu.InputFields
+					withLabels
+					format="hexa"
+				/>
 
 				<span
 					style={{
@@ -64,7 +66,7 @@ const ExampleColorPicker = ({ color }: { color: string }) => {
 					Final Color : {value}
 				</span>
 			</div>
-		</Rangu.Root>
+		</Rangu>
 	);
 };