+++ src/current-color/index.tsx
@@ -1,4 +1,3 @@
-import { useRangu } from "@/hooks/use-rangu";
 import { cn } from "@/utils";
 import * as React from "react";
 import { ColorSwatch, type ColorSwatchProps } from "react-aria-components";
@@ -10,12 +9,9 @@ const RanguCurrentColor = React.forwardRef<
 	HTMLDivElement,
 	RanguCurrentColorProps
 >(({ className, ...rest }, forwardedRef) => {
-	const { value } = useRangu();
-
 	return (
 		<ColorSwatch
 			{...rest}
-			color={value}
 			ref={forwardedRef}
 			className={cn(
 				"size-14 rounded-lg shadow-sm shadow-slate-200",